+++ platform-tracing-core/src/main/java/space/br1440/platform/tracing/core/manual/OperationSpanBuilderImpl.java
@@ -1,20 +1,20 @@
 package space.br1440.platform.tracing.core.manual;
 
 import jakarta.annotation.Nonnull;
-import space.br1440.platform.tracing.api.span.builder.OperationSpanBuilder;
+import space.br1440.platform.tracing.api.propagation.OtelTraceparentReader;
 import space.br1440.platform.tracing.api.span.SpanCategory;
+import space.br1440.platform.tracing.api.span.builder.OperationSpanBuilder;
 import space.br1440.platform.tracing.core.runtime.TracingRuntime;
 import space.br1440.platform.tracing.core.semconv.policy.AttributePolicy;
-
-import java.util.Objects;
 
 final class OperationSpanBuilderImpl extends AbstractSemanticSpanBuilder<OperationSpanBuilder>
         implements OperationSpanBuilder {
 
     OperationSpanBuilderImpl(@Nonnull TracingRuntime implementation,
                              @Nonnull AttributePolicy policy,
+                             @Nonnull OtelTraceparentReader traceparentReader,
                              @Nonnull String name) {
-        super(implementation, policy, SpanCategory.INTERNAL, name, "OperationSpanBuilder");
+        super(implementation, policy, traceparentReader, SpanCategory.INTERNAL, name, "OperationSpanBuilder");
     }
 
     @Override